requirements.txt
certifi==2017.11.5
chardet==3.0.4
gTTS==1.2.2
gTTS-token==1.1.1
idna==2.6
mpg123==0.4
PyAudio==0.2.11
pytz==2017.3
requests==2.18.4
SpeechRecognition==3.8.1
urllib3==1.22
beautifulsoup4==4.6.0
weather-api==0.0.4